requirements.txt
fastapi
uvicorn
ultralytics
Pillow
PyYAML
python-multipart
numpy
opencv-python
torch
torchvision
tensorflow
requests